fix: several typos & mis-uses

+++ main.py
@@ -26,7 +26,7 @@ METADATA_URL = 'http://metadata.google.internal/computeMetadata/v1/'
 METADATA_HEADERS = {'Metadata-Flavor': 'Google'}
 
 # ROLE: tidb, tikv, pd
-ROLE: str = os.environ["ROLE"]
+ROLE: str = os.environ["ROLE"].lower()
 TC_NAME: str = os.environ["CLUSTER_NAME"]
 TLS: str = os.environ["TLS"] if os.getenv("TLS") else "false"
 TIKV_TLS: str = "" if TLS == "false" else "--ca-path /var/lib/tikv-tls/ca.crt --cert-path /var/lib/tikv-tls/tls.crt --key-path /var/lib/tikv-tls/tls.key"
@@ -37,9 +37,10 @@ PD_ADDR: str = f"{HTTP_PREFIX}{TC_NAME}-pd:2379"
 # DEBUG: bool = os.environ.get("CLUSTER_NAME") is not None
 DEBUG: bool = True  # FIXME
 
-#Init the Kubernetes client
+# Init the Kubernetes client
 config.load_incluster_config()
 api = client.CoreV1Api()
+
 
 def wait_for_maintenance():
     url = METADATA_URL + 'instance/maintenance-event'
@@ -58,7 +59,7 @@ def wait_for_maintenance():
                 headers=METADATA_HEADERS)
         except requests.exceptions.ConnectionError:
             # The metadata server may not be available, so retry.
-            continue    
+            continue
 
         # During maintenance the service can return a 503, so these should
         # be retried.
@@ -80,7 +81,7 @@ def wait_for_maintenance():
             elif ROLE == "tikv":
                 evict_store()
                 store_evicted = True
-            else:
+            elif ROLE == "tidb":
                 schedule_tidb_node(True)
                 delete_tidb_pod()
         elif is_during_maintenance(maintenance_event, last_maintenance_event):
@@ -104,27 +105,27 @@ def is_during_maintenance(maintenance_event, last_maintenance_event) -> bool:
     return (maintenance_event is not None and
             maintenance_event == last_maintenance_event)
 
-def schedule_tidb_node(cordon_node: bool):
-    node_name = get_hostname()
-    node = api.read_node_status(node_name)
-    # Update the node's metadata.
-    node.spec.unschedulable = cordon_node
-    api.patch_node_status(node_name, node)
 
-def evict_tidb():
-    namespace = get_namespace()
-    pod = get_hostname()
-    patch = {"spec": {"containers": [{"name": "tidb", "image": "not-exist-image"}]}}
-    api.patch_namespaced_pod(name=pod, namespace=namespace, body=patch)
+def schedule_tidb_node(cordon_node: bool):
+    node_name = get_self_nodename()
+    body = {
+        "spec": {
+            "unschedulable": cordon_node,
+        },
+    }
+    api.patch_node(node_name, body)
+
 
 def delete_tidb_pod():
     namespace = get_namespace()
     pod = get_hostname()
     api.delete_namespaced_pod(name=pod, namespace=namespace)
 
+
 def get_namespace() -> str:
     with open('/var/run/secrets/kubernetes.io/serviceaccount/namespace', 'r') as f:
         return f.read().strip()
+
 
 def resign_leader():
     leader = get_leader()
@@ -137,11 +138,18 @@ def resign_leader():
         print(f"resigning pd leader, cmd [{cmd}]")
         shell_cmd(cmd)
 
+
 def get_leader() -> str:
-    return shell_cmd(f"pd-ctl member leader show --pd {PD_ADDR} {PD_TLS} | grep 'name' | cut -d: -f2").strip().strip(',').strip('"')
+    return shell_cmd(f"/pd-ctl member leader show --pd {PD_ADDR} {PD_TLS} | grep 'name' | cut -d: -f2").strip().strip(',').strip('"')
+
 
 def get_hostname() -> str:
-    return shell_cmd(f"hostname").strip()
+    return shell_cmd("hostname").strip()
+
+
+def get_self_nodename() -> str:
+    return os.environ["NODENAME"]
+
 
 def evict_store():
     store_id = get_store_id()
@@ -158,7 +166,7 @@ def recover_restore():
 
 
 def get_store_id() -> str:
-    return shell_cmd("/tikv-ctl --host 127.0.0.1:20160 {TIKV_TLS} | grep 'store id' | cut -d: -f2").strip()
+    return shell_cmd(f"/tikv-ctl --host 127.0.0.1:20160 {TIKV_TLS} store | grep 'store id' | cut -d: -f2").strip()
 
 
 def shell_cmd(cmd: str) -> str:
@@ -171,7 +179,7 @@ def shell_cmd(cmd: str) -> str:
             "utf8"), stderr.decode("utf8"), cmd)
 
     if DEBUG:
-        print(f'stdout: {stdout.decode("utf8").strip()}, stderr: {stderr.decode("utf8").strip()}')
+        print(f'cmd: {cmd}, stdout: {stdout.decode("utf8").strip()}, stderr: {stderr.decode("utf8").strip()}')
 
     return stdout.decode("utf8")
